ResInsight: Linux compile fixes p4#: 18783

+++ ApplicationCode/UserInterface/RiuSimpleHistogramWidget.h
@@ -21,7 +21,7 @@ private:
     int                 xPosFromColIdx(size_t colIdx)          { return  (int)(m_x + 1 + (m_width - 2 ) * colIdx/m_histogramData.size());}
     int                 yPosFromCount(size_t colHeight)        { return  (int)(m_y + m_height - 1 - (m_height - 3 ) * colHeight/m_maxHistogramCount);}
 
-    int                 xPosFromDomainValue(double value) { double range = m_max - m_min; return  (range == 0.0) ? (m_x + 1) : (int)(m_x + 1 + (m_width - 2 ) * (value - m_min)/(m_max - m_min));}
+    int                 xPosFromDomainValue(double value) { double range = m_max - m_min; return  (range == 0.0) ? (int)(m_x + 1) : (int)(m_x + 1 + (m_width - 2 ) * (value - m_min)/(m_max - m_min));}
 
     std::vector<size_t> m_histogramData;
     double              m_max;
